requirements.txt
aio-pika==9.4.3
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiormq==6.8.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
DateTime==5.5
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
faiss-cpu==1.9.0
fastapi==0.115.4
fonttools==4.55.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.4
langchain-core==0.3.19
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.143
loguru==0.7.2
marshmallow==3.23.1
matplotlib==3.9.2
motor==3.6.0
multidict==6.1.0
mypy-extensions==1.0.0
nltk==3.9.1
numpy==1.26.4
openai==1.54.4
orjson==3.10.11
packaging==24.2
pamqp==3.3.0
pgvector==0.3.6
pika==1.3.2
pika-stubs==0.1.3
pillow==11.0.0
pluggy==1.5.0
propcache==0.2.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg_pool==3.2.6
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pymongo==4.9.2
redis[hiredis]==5.0.1
pyparsing==3.2.0
pyphen==0.17.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
textblob==0.18.0.post0
textstat==0.7.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
win32-setctime==1.1.0
yarl==1.17.2
zope.interface==7.1.1
datadog-api-client==2.32.0